github - updating my own github event filter class
Unfortunately the "my events" github api endpoint does not give any info
on if the repository is private or public. I wanted to deactivate the
display of events coming from a public repo, because Slack will resolve
the url and provide a summary of the page behind the url, which makes
the bot output less readable. I don't want to make an extra http query
to figure it out.

So let's go for a static list of blocklisted repositories instead.